requirements.txt
pandas>=1.5.0
tweepy>=4.12.0
matplotlib>=3.5.0
seaborn>=0.12.0
python-dotenv>=0.19.0 